# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/users/profile": {
      "get": {
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "operationId": "UsersController_getAllUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/devices": {
      "post": {
        "operationId": "DevicesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      },
      "get": {
        "operationId": "DevicesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/devices/{id}": {
      "get": {
        "operationId": "DevicesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "put": {
        "operationId": "DevicesController_updateDevice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDeviceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "operationId": "DevicesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/devices/command": {
      "post": {
        "operationId": "DevicesController_sendCommand",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceCommandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/pumps": {
      "post": {
        "operationId": "PumpsController_create",
        "summary": "Create a new pump",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePumpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Pumps"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "PumpsController_findAll",
        "summary": "Get all pumps",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Pumps"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/pumps/{id}": {
      "get": {
        "operationId": "PumpsController_findOne",
        "summary": "Get pump by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Pumps"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "PumpsController_remove",
        "summary": "Delete pump",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Pumps"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/pumps/{id}/toggle": {
      "put": {
        "operationId": "PumpsController_togglePump",
        "summary": "Toggle pump on/off",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Pumps"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/pumps/swap": {
      "post": {
        "operationId": "PumpsController_swapPumps",
        "summary": "Swap two pumps",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwapPumpsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Pumps"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/pumps/{id}/status": {
      "put": {
        "operationId": "PumpsController_updateStatus",
        "summary": "Update pump status (water level, wifi, etc.)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Pumps"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/pumps/failover": {
      "post": {
        "operationId": "PumpsController_handleFailover",
        "summary": "Trigger failover (activate backup pump)",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Pumps"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/pumps/{id}/command": {
      "post": {
        "operationId": "PumpsController_sendCommand",
        "summary": "Send command to pump (motor, height, timer)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Pumps"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/pumps/{id}/data": {
      "get": {
        "operationId": "PumpsController_getPumpData",
        "summary": "Get pump real-time data",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Pumps"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/monitoring/usage": {
      "post": {
        "operationId": "MonitoringController_recordUsage",
        "summary": "Record usage data",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Monitoring"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/monitoring/usage/daily/{pumpId}": {
      "get": {
        "operationId": "MonitoringController_getDailyUsage",
        "summary": "Get daily usage records",
        "parameters": [
          {
            "name": "pumpId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Monitoring"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/monitoring/usage/monthly/{pumpId}": {
      "get": {
        "operationId": "MonitoringController_getMonthlyUsage",
        "summary": "Get monthly usage records",
        "parameters": [
          {
            "name": "pumpId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "months",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Monitoring"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/monitoring/usage/total": {
      "get": {
        "operationId": "MonitoringController_getTotalUsage",
        "summary": "Get total usage for user",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Monitoring"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "MQTT Device API",
    "description": "API documentation for MQTT Device Management",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "Add your JWT token as: Bearer <token>"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {}
      },
      "LoginDto": {
        "type": "object",
        "properties": {}
      },
      "CreateDeviceDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateDeviceDto": {
        "type": "object",
        "properties": {}
      },
      "DeviceCommandDto": {
        "type": "object",
        "properties": {}
      },
      "CreatePumpDto": {
        "type": "object",
        "properties": {}
      },
      "SwapPumpsDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}